requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
cffi==1.15.1
click==8.1.3
comm==0.1.3
cryptography==40.0.2
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
haikunator==2.1.0
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyyaml==6.0
pycparser==2.21
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.0.2
six==1.16.0
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
wcwidth==0.2.6
